requirements.txt
torch
pandas
numpy
scikit-learn
scipy
groq
python-dotenv
rdkit
optuna
